Move doctest to test folder
Enable test executable compilation
