+++ packages/drqp_serial/test/CMakeLists.txt
@@ -1,20 +1,20 @@
-cmake_minimum_required(VERSION 3.5)
 
-set(PROJECT_NAME TestDrQP)
-project(${PROJECT_NAME})
-
-file(GLOB_RECURSE ALL_CODE_HDRS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
-file(GLOB_RECURSE ALL_CODE_SRCS ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
-
+find_package(catch_ros2 REQUIRED)
 find_package(Boost REQUIRED COMPONENTS thread serialization system)
 
-add_executable(${PROJECT_NAME}
-  ${ALL_CODE_HDRS}
-  ${ALL_CODE_SRCS}
-  ${SOURCE_FILES})
-
-target_link_libraries(${PROJECT_NAME} PRIVATE servo Boost::thread Boost::serialization Boost::system)
-
-file(INSTALL ${CMAKE_CURRENT_SOURCE_DIR}/test_data
-  DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/tests
+add_executable(drqp_serial_test_catch
+  drqp_serial_test.cpp
 )
+
+# https://docs.ros.org/en/humble/p/catch_ros2/index.html
+target_link_libraries(drqp_serial_test_catch PRIVATE
+  catch_ros2::catch_ros2_with_main
+  drqp_serial
+)
+
+add_executable(drqp_serial_tests
+  TestSerialRecordingProxy.cpp
+  TestsMain.cpp
+)
+
+target_link_libraries(drqp_serial_tests PRIVATE drqp_serial Boost::thread Boost::serialization Boost::system)
