+++ packages/drqp_serial/test/TestSerialRecordingProxy.cpp
@@ -23,7 +23,7 @@
 #include "drqp_serial/SerialPlayer.h"
 #include "drqp_serial/SerialRecordingProxy.h"
 #include "drqp_serial/UnixSerial.h"
-#include "drqp_serial/doctest.h"
+#include "doctest.h"
 
 void simpleSerialTest(SerialProtocol& serial)
 {
